requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
google-auth==2.0.1
google-auth-oauthlib==0.4.5
gspread==4.0.1
idna==3.2
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
youtube-dl==2021.6.6
